requirements.txt
#nsml: nsml/ml:cuda10.1-cudnn7-pytorch1.3keras2.3
pandas~=1.1.3
numpy
pymongo~=3.9.0
fire~=0.2.1
pillow~=6.2.1
torch_optimizer
randaugment
easydict
